random agent populates the orderbook now, gnuplot also working, although tested only on fixed srand seed

+++ include/order.h
@@ -16,7 +16,7 @@ typedef enum {
 } order_type;
 
 typedef struct order {
-    uint64_t   id;
+    uint32_t   id;
     uint32_t   price;
     uint32_t   qty;
     order_side side;
